chore: module API attempt 2
