+++ src/index.ts
@@ -1,1 +1,3 @@
-export * from "./lib";
+export * from "./swig";
+export * as swig from "./swig";
+export * as emscripten from "./emscripten";
