+++ test/unit/lib/threads/runnable.js
@@ -1,4 +1,5 @@
-const bls = require("../../../../dist/lib");
+require("reflect-metadata");
+const bls = require("../../../../dist/index");
 
 const msg = Buffer.from("sample-msg");
