+++ tsconfig.json
@@ -21,6 +21,9 @@
     "declaration": true,
     "sourceMap": true,
     "declarationMap": true,
-    "esModuleInterop": true
+    "esModuleInterop": true,
+
+    "experimentalDecorators": true,
+    "emitDecoratorMetadata": true
   }
 }
